requirements.txt
scikit-learn
git+https://github.com/huggingface/transformers
git+https://github.com/huggingface/datasets
evaluate
pyyaml
sagemaker
gensim
"dill<0.3.5"
nvidia-ml-py3
nltk
gevent
peewee
pymysql
transformers[sentencepiece]
redlines